+++ ADSOrcid/app.py
@@ -1,365 +1,748 @@
-from __future__ import absolute_import, unicode_literals
+
+
+from .models import ClaimsLog, Records, AuthorInfo, ChangeLog
+from .utils import get_date, setup_logging
+from ADSOrcid import utils, names
+from ADSOrcid.exceptions import IgnorableException
 from celery import Celery
-from celery.utils.log import get_task_logger
-from celery import Task
-import traceback
-from kombu import Exchange, Queue
-import math
-
-from ADSOrcid import db, utils
-from ADSOrcid.exceptions import ProcessingException, IgnorableException
-from ADSOrcid import matcher, updater, importer
-from ADSOrcid.models import KeyValue
+from contextlib import contextmanager
+from dateutil.tz import tzutc
+from sqlalchemy import and_
+from sqlalchemy import create_engine
+from sqlalchemy.orm import scoped_session
+from sqlalchemy.orm import sessionmaker
+import cachetools
 import datetime
+import json
+import os
+import random
 import requests
+import time
+import traceback
 
-logger = utils.setup_logging('app', 'ADSOrcid')
-conf = utils.load_config()
-app = Celery('ADSOrcid',
+
+# global objects; we could make them belong to the app object but it doesn't seem necessary
+# unless two apps with a different endpint/config live along; XXX - move if necessary
+cache = cachetools.TTLCache(maxsize=1024, ttl=3600, timer=time.time, missing=None, getsizeof=None)
+orcid_cache = cachetools.TTLCache(maxsize=1024, ttl=3600, timer=time.time, missing=None, getsizeof=None)
+ads_cache = cachetools.TTLCache(maxsize=1024, ttl=3600, timer=time.time, missing=None, getsizeof=None)
+bibcode_cache = cachetools.TTLCache(maxsize=2048, ttl=3600, timer=time.time, missing=None, getsizeof=None)
+
+ALLOWED_STATUS = set(['claimed', 'updated', 'removed', 'unchanged', 'forced', '#full-import'])
+
+
+def create_app(app_name='ADSOrcid',
+               local_config=None):
+    """Builds and initializes the Celery application."""
+    
+    conf = utils.load_config()
+    if local_config:
+        conf.update(local_config)
+
+    app = ADSOrcidCelery(app_name,
              broker=conf.get('CELERY_BROKER', 'pyamqp://'),
-             include=['ADSOrcid.app'])
+             include=conf.get('CELERY_INCLUDE', ['ADSOrcid.app']))
 
-
-exch = Exchange(conf.get('CELERY_DEFAULT_EXCHANGE', 'ads-orcid'), type=conf.get('CELERY_DEFAULT_EXCHANGE_TYPE', 'topic'))
-
-app.conf.CELERY_QUEUES = (
-    Queue('errors', exch, routing_key='errors', durable=False, message_ttl=24*3600*5),
-    Queue('new-claim', exch, routing_key='new-claim'),
-    Queue('verified-claim', exch, routing_key='verified-claim'),
-    Queue('output', exch, routing_key='output'),
-)
-
-app.conf.update(conf)
-db.init_app()
-
-class MyTask(Task):
-    def on_failure(self, exc, task_id, args, kwargs, einfo):
-        logger.error('{0!r} failed: {1!r}'.format(task_id, exc))
+    app.init_app(conf)
+    return app
 
 
 
-@app.task(base=MyTask, queue='check-orcidid')
-def task_index_orcid_profile(message):
-    """
-    Fetch a fresh profile from the orcid-service and compare
-    it against the state of the storage (diff). And re-index/update
-    them.
+class ADSOrcidCelery(Celery):
     
-
-    :param message: contains the message inside the packet
-        {
-         'orcidid': '.....',
-         'start': 'ISO8801 formatted date (optional), indicates 
-             the moment we checked the orcid-service'
-         'force': Boolean (if present, we'll not skip unchanged
-             profile)
-        }
-    :return: no return
-    """
-    
-    if 'orcidid' not in message:
-        raise IgnorableException('Received garbage: {}'.format(message))
-    
-    message['start'] = utils.get_date()
-    orcidid = message['orcidid']
-
-    orcid_present, updated, removed = importer.get_claims(orcidid,
-                         app.conf.get('API_TOKEN'), 
-                         app.conf.get('API_ORCID_EXPORT_PROFILE') % orcidid,
-                         force=message.get('force', False),
-                         orcid_identifiers_order=app.conf.get('ORCID_IDENTIFIERS_ORDER', {'bibcode': 9, '*': -1})
-                         )
-    
-    if not updated and not removed:
-        # reschedule re-check
-        task_index_orcid_profile.apply_async(message, countdown = app.conf.get('ORCID_PROFILE_RECHECK_WINDOW', 3600*24))
-        return
-    
-    to_claim = []
-    
-    #always insert a record that marks the beginning of a full-import
-    #TODO: record orcid's last-modified-date
-    to_claim.append(importer.create_claim(bibcode='', 
-                                              orcidid=orcidid, 
-                                              provenance='OrcidImporter', 
-                                              status='#full-import',
-                                              date=utils.get_date()
-                                              ))
-    
-    # find difference between what we have and what orcid has
-    claims_we_have = set(updated.keys()).difference(set(removed.keys()))
-    claims_orcid_has = set(orcid_present.keys())
-    
-    # those guys will be added (with ORCID date signature)
-    for c in claims_orcid_has.difference(claims_we_have):
-        claim = orcid_present[c]
-        to_claim.append(importer.create_claim(bibcode=claim[0], 
-                                              orcidid=orcidid, 
-                                              provenance=claim[2], 
-                                              status='claimed', 
-                                              date=claim[1])
-                                              )
-    
-    # those guys will be removed (since orcid doesn't have them)
-    for c in claims_we_have.difference(claims_orcid_has):
-        claim = updated[c]
-        to_claim.append(importer.create_claim(bibcode=claim[0], 
-                                              orcidid=orcidid, 
-                                              provenance='OrcidImporter', 
-                                              status='removed')
-                                              )
+    def __init__(self, app_name, *args, **kwargs):
+        Celery.__init__(self, *args, **kwargs)
+        self._config = utils.load_config()
+        self._session = None
+        self.logger = utils.setup_logging(app_name, app_name) #default logger
         
-    # and those guys will be updated if their creation date is significantly off
-    for c in claims_orcid_has.intersection(claims_we_have):
+    
+
+    def init_app(self, config=None):
+        """This function must be called before you start working with the application
+        (or worker, script etc)
         
-        orcid_claim = orcid_present[c]
-        ads_claim = updated[c]
+        :return None
+        """
         
-        delta = orcid_claim[1] - ads_claim[1]
-        if delta.total_seconds() > app.conf.get('ORCID_UPDATE_WINDOW', 60): 
-            to_claim.append(importer.create_claim(bibcode=orcid_claim[0], 
-                                              orcidid=orcidid, 
-                                              provenance='OrcidImporter', 
-                                              status='updated',
-                                              date=orcid_claim[1])
-                                              )
-        elif message.get('force', False):
-            to_claim.append(importer.create_claim(bibcode=orcid_claim[0], 
-                                              orcidid=orcidid, 
-                                              provenance='OrcidImporter', 
-                                              status='forced',
-                                              date=orcid_claim[1])
-                                              )
+        if self._session is not None: # the app was already instantiated
+            return
+        
+        if config:
+            self._config.update(config)
+        
+        self.logger = utils.setup_logging(__file__, 'app', self._config.get('LOGGING_LEVEL', 'INFO'))
+        self._engine = create_engine(config.get('SQLALCHEMY_URL', 'sqlite:///'),
+                               echo=config.get('SQLALCHEMY_ECHO', False))
+        self._session_factory = sessionmaker()
+        self._session = scoped_session(self._session_factory)
+        self._session.configure(bind=self._engine)
+    
+    
+    def close_app(self):
+        """Closes the app"""
+        self._session = self._engine = self._session_factory = None
+        self.logger = None
+    
+        
+    @contextmanager
+    def session_scope(self):
+        """Provides a transactional session - ie. the session for the 
+        current thread/work of unit.
+        
+        Use as:
+        
+            with session_scope() as session:
+                o = AuthorInfo(...)
+                session.add(o)
+        """
+    
+        if self._session is None:
+            raise Exception('init_app() must be called before you can use the session')
+        
+        # create local session (optional step)
+        s = self._session()
+        
+        try:
+            yield s
+            s.commit()
+        except:
+            s.rollback()
+            raise
+        finally:
+            s.close()  
+            
+    
+    def insert_claims(self, claims):
+        """
+        Build a batch of claims and saves them into a database
+        
+        :param: claims - list of json values, with claims
+                       - or list of claims (ClaimLog) instances
+        :return number of claims that were successfuly added
+                to the database
+        """
+        res = []
+        with self.session_scope() as session:
+            for c in claims:
+                if isinstance(c, ClaimsLog):
+                    claim = c
+                else:
+                    claim = self.create_claim(**c)
+                if claim:
+                    session.add(claim)
+                    res.append(claim)
+            session.commit()
+            res = [x.toJSON() for x in res]
+        return res
+    
+    def create_claim(self, 
+                 bibcode=None, 
+                 orcidid=None, 
+                 provenance=None, 
+                 status=None, 
+                 date=None, 
+                 force_new=True,
+                 **kwargs):
+        """
+        Inserts (or updates) ClaimLog entry.
+        
+        :return: ClaimsLog instance (however this is only for reading, you should
+            not try to do anything with it; the session will have been closed already)
+        """
+        assert(orcidid)
+        if isinstance(date, basestring):
+            date = utils.get_date(date)
+        if status and status.lower() not in ALLOWED_STATUS:
+            raise Exception('Unknown status %s' % status)
+        
+        if not date or force_new is True: # we don't need to verify the record exists
+            return ClaimsLog(bibcode=bibcode, 
+                      orcidid=orcidid,
+                      provenance=provenance, 
+                      status=status,
+                      created=date or utils.get_date())
         else:
-            to_claim.append(importer.create_claim(bibcode=orcid_claim[0], 
-                                              orcidid=orcidid, 
-                                              provenance='OrcidImporter', 
-                                              status='unchanged',
-                                              date=orcid_claim[1]))
+            with self.session_scope() as session:
+                f = session.query(ClaimsLog).filter_by(created=date).first()
+                if f and f.bibcode == bibcode and f.orcidid == orcidid:
+                    f.provenance = provenance
+                    f.status = status
+                else:
+                    return ClaimsLog(bibcode=bibcode, 
+                      orcidid=orcidid,
+                      provenance=provenance, 
+                      status=status,
+                      created=date)
 
-    if len(to_claim):
-        # create record in the database
-        json_claims = importer.insert_claims(to_claim)
-        # set to the queue for processing
-        for claim in json_claims:
-            claim['bibcode_verified'] = True
-            task_ingest_claim.delay(claim)
+
+    def import_recs(self, input_file, default_provenance=None, 
+                default_status='claimed', collector=None):
+        """
+        Imports (creates log records) of claims from
+        :param: input_file - String, path to the file with the following 
+                information (tab delimited):
+                    bibcode
+                    orcid_id
+                    provenance - optional
+                    status - optional
+                    date - optional
+        :param: default_provenance - String, this will be used if the records
+                don't provide provenance
+        :param: default_status - String, used when status is not supplied
+        :param: collector - if passed in, the results will be inserted
+                into it
+        :type: array
+        """
+        
+        if not os.path.exists(input_file):
+            raise Exception('{file} does not exist'.format(
+                               file=input_file
+                               ))
+        if collector is not None:
+            assert(isinstance(collector, list))
             
-    # reschedule future check
-    task_index_orcid_profile.apply_async(message, countdown = app.conf.get('ORCID_PROFILE_RECHECK_WINDOW', 3600*24))
-
-
-
-@app.task(base=MyTask, queue='record-claim')
-def task_ingest_claim(msg, **kwargs):
-    """
-    Processes claims in the system; it enhances the claim
-    with the information about the claimer. (and in the
-    process, updates our knowledge about the ORCIDID).
-    
-    Results are published into the queue 'verified-claim'
-    
-    :param msg: contains the message inside the packet
-        {'bibcode': '....',
-        'orcidid': '.....',
-        'provenance': 'string (optional)',
-        'status': 'claimed|updated|deleted (optional)',
-        'date': 'ISO8801 formatted date (optional)'
-        }
-    :return: no return
-    """
-    
-    if not isinstance(msg, dict):
-        raise ProcessingException('Received unknown payload {0}'.format(msg))
-    
-    if not msg.get('orcidid'):
-        raise ProcessingException('Unusable payload, missing orcidid {0}'.format(msg))
-
-    if msg.get('status', 'created') in ('unchanged', '#full-import'):
-        return
-                    
-    author = matcher.retrieve_orcid(msg['orcidid'])
-    
-    if not author:
-        raise ProcessingException('Unable to retrieve info for {0}'.format(msg['orcidid']))
-    
-    # clean up the bicode
-    bibcode = msg['bibcode'].strip()
-    
-    if not msg.get('bibcode_verified', False):
-        if ' ' in bibcode:
-            parts = bibcode.split()
-            l = [len(x) for x in parts]
-            if 19 in l:
-                bibcode = parts[l.index(19)] 
-        
-        # check if we can translate the bibcode/identifier
-        rec = updater.retrieve_metadata(bibcode)
-        if rec.get('bibcode') != bibcode:
-            logger.warning('Resolving {0} into {1}'.format(bibcode, rec.get('bibcode')))
-        bibcode = rec.get('bibcode') 
-    
-    msg['bibcode'] = bibcode
-    msg['name'] = author['name']
-    if author.get('facts', None):
-        for k, v in author['facts'].iteritems():
-            msg[k] = v
+        if default_provenance is None:
+            default_provenance = os.path.abspath(input_file)
             
-    msg['author_status'] = author['status']
-    msg['account_id'] = author['account_id']
-    msg['author_updated'] = author['updated']
-    msg['author_id'] = author['id']
-    
-    if msg['author_status'] in ('blacklisted', 'postponed'):
-        return
-    
-    task_match_claim.delay(msg)
+        def rec_builder(bibcode=None, orcidid=None, provenance=None, status=None, date=None):
+            assert(bibcode and orcidid)
+            return ClaimsLog(bibcode=bibcode, 
+                          orcidid=orcidid,
+                          provenance=provenance or default_provenance, 
+                          status=status or default_status,
+                          created=date and get_date(date) or get_date())
+            
+        i = 0
+        with open(input_file, 'r') as fi:
+            with self.session_scope() as session:
+                for line in fi:
+                    i += 1
+                    l = line.strip()
+                    if len(l) == 0 or l[0] == '#':
+                        continue
+                    parts = l.split('\t')
+                    try:
+                        rec = rec_builder(*parts)
+                        session.add(rec)
+                        if collector is not None:
+                            collector.append(rec.toJSON())
+                    except Exception, e:
+                        self.logger.error('Error importing line %s (%s) - %s' % (i, l, e))
+                    if i % 1000 == 0:
+                        session.commit()
+                session.commit()
 
 
-
-@app.task(base=MyTask, queue='match-claim')
-def task_match_claim(claim, **kwargs):
-    """
-    Takes the claim, matches it in the database (will create
-    entry for the record, if not existing yet) and updates 
-    the metadata.
-    
-    :param claim: contains the message inside the packet
-        {'bibcode': '....',
-        'orcidid': '.....',
-        'name': 'author name',
-        'facts': 'author name variants',
-        }
-    :return: no return
-    """
-    
-    if not isinstance(claim, dict):
-        raise ProcessingException('Received unknown payload {0}'.format(claim))
-    
-    if not claim.get('orcidid'):
-        raise ProcessingException('Unusable payload, missing orcidid {0}'.format(claim))
-
-    bibcode = claim['bibcode']
-    rec = updater.retrieve_record(bibcode)
-    
-    
-    cl = updater.update_record(rec, claim)
-    if cl:
-        updater.record_claims(bibcode, rec['claims'], rec['authors'])
-        # TODO: call directly? circumvent the queue?
-        task_output_results.delay({'authors': rec.get('authors'), 'bibcode': rec['bibcode'], 'claims': rec.get('claims')})
-    else:
-        logger.warning('Claim refused for bibcode:{0} and orcidid:{1}'
-                        .format(claim['bibcode'], claim['orcidid']))
-
-
-@app.task(base=MyTask, queue='output-results')
-def task_output_results(producer, msg):
-    """
-    This worker will forward results to the outside 
-    exchange (typically an ADSImportPipeline) to be
-    incorporated into the storage
-    """
-    Task.apply_async(args=('orcid_claims', msg),
-                     exchange=app.conf.get('OUTPUT_EXCHANGE', 'import-pipeline'),
-                     queue=app.conf.get('OUTPUT_QUEUE', 'update-record'),
-                     routing_key=app.conf.get('OUTPUT_QUEUE', 'update-record'))
-
-
-
-@app.task(base=MyTask, queue='errors')
-def task_handle_errors(producer, message, redelivered=False):
-    """
-    @param producer: string, the name of the queue where the error originated
-    @param body: 
-    """
-    
-    logger.error(u'\nproducer={}\nmsg={}\n'.format(producer, message))
-    
-    
-
-
-@app.task(base=MyTask, queue='check-updates')
-def task_check_orcid_profiles(msg):
-    """Check the orcid microservice for updated orcid profiles.
-    
-    This function is somewhat complex
-    we are trying to defend against multiple executions (assuming 
-    that there is many workers and each of them can receive its own
-    signal to start processing). 
-    
-    Basically, we'll only want to check for updated profiles once.
-    And the synchronization is done via a database. So the worker
-    must update the 'last.check' timestamp immediately (and we
-    'optimistically' hope that it will be enough to prevent clashes;
-    well - even if that is not a strong guarantee, it wouldn't be 
-    a tragedy if a profile is checked twice...)
-    
-    Additional difficulty is time synchronization: the worker can 
-    be executed as often as you like, but it will refuse to do any
-    work unless the time window between the checks is large enough.
-    """
-    
-    with db.session_scope() as session:
-        kv = session.query(KeyValue).filter_by(key='last.check').first()
-        if kv is None:
-            kv = KeyValue(key='last.check', value='1974-11-09T22:56:52.518001Z') #force update
-        
-        latest_point = utils.get_date(kv.value) # RFC 3339 format
-        now = utils.get_date()
-        
-        total_wait = app.conf.get('ORCID_CHECK_FOR_CHANGES', 60*5) #default is 5min
-        delta = now - latest_point
-        
-        if delta.total_seconds() < total_wait:
-            # register our own execution in the future
-            task_check_orcid_profiles.apply_async(msg, countdown=(total_wait - delta.total_seconds()) + 1)
+    def _get_ads_orcid_profile(self, orcidid, api_token, api_url):
+        r = requests.get(api_url,
+                 params={'reload': True},
+                 headers={'Accept': 'application/json', 'Authorization': 'Bearer:%s' % api_token})
+        if r.status_code == 200:
+            return r.json()
         else:
-            logger.info("Checking for orcid updates")
-            
-            # increase the timestamp by one microsec and get new updates
-            latest_point = latest_point + datetime.timedelta(microseconds=1)
-            r = requests.get(app.conf.get('API_ORCID_UPDATES_ENDPOINT') % latest_point.isoformat(),
-                        params={'fields': ['orcid_id', 'updated', 'created']},
-                        headers = {'Authorization': 'Bearer {0}'.format(app.conf.get('API_TOKEN'))})
-            
-            if r.status_code != 200:
-                logger.error('Failed getting {0}\n{1}'.format(
-                            app.conf.get('API_ORCID_UPDATES_ENDPOINT') % kv.value,
-                            r.text))
-                msg['errcount'] = msg.get('errcount', 0) + 1
-                
-                # schedule future execution offset by number of errors (rca: do exponential?)
-                task_check_orcid_profiles.apply_async(msg, countdown = total_wait * msg['errcount'])
+            self.logger.warning('Missing profile for: {0}'.format(orcidid))
+            self.logger.warning(r.text)
+            return {}
+
+    def get_claims(self, orcidid, api_token, api_url, force=False, 
+                      orcid_identifiers_order=None):
+        """
+        Fetch a fresh profile from the orcid-service and compare
+        it against the state of the storage (diff). Return the docs
+        that need updating, and removal
+        
+    
+        :param orcidid
+            - string, orcid identifier
+        :param: api_token
+            - string, OAuth token to access ADS API
+        :param: api_url:
+            - string, URL for getting the orcid profiles
+        :param: force
+            - bool, when True it forces claims to be counted
+                as new (even if we have already indexed them)
+        :param: orcid_identifiers_order
+            - dict, helps to sort claims by their identifies.
+                (e.g. to say that bibcodes have higher priority than
+                dois)
+        :return: 
+            - updated: dict of bibcodes that were updated
+                - keys are lowercased bibcodes
+                - values are (bibcode, timestamp)
+            - removed: set of bibcodes that were removed
+                - keys are lowercased bibcodes
+                - values are (bibcode, timestamp)
+        """
+        
+        
+        
+    
+        # make sure the author is there (even if without documents) 
+        author = self.retrieve_orcid(orcidid) # @UnusedVariable
+        data = self._get_ads_orcid_profile(orcidid, api_token, api_url)
+        
+        if data is None:
+            return #TODO: remove all existing claims?
+        
+        profile = data.get('profile', {})
+        if not profile:
+            return #TODO: remove all existing claims?
+        
+    
+        with self.session_scope() as session:
+              
+            # orcid is THE ugliest datastructure of today!
+            try:
+                works = profile['orcid-profile']['orcid-activities']['orcid-works']['orcid-work']
+            except:
+                self.logger.warning('Nothing to do for: '
+                    '{0} ({1})'.format(orcidid,
+                                       traceback.format_exc()))
                 return
+    
+            # check we haven't seen this very profile already
+            try:
+                updt = str(profile['orcid-profile']['orcid-history']['last-modified-date']['value'])
+                updt = float('%s.%s' % (updt[0:10], updt[10:]))
+                updt = datetime.datetime.fromtimestamp(updt, tzutc())
+                updt = get_date(updt.isoformat())
+            except KeyError:
+                updt = get_date()
+                                    
+            # find the most recent #full-import record
+            last_update = session.query(ClaimsLog).filter(
+                and_(ClaimsLog.status == '#full-import', ClaimsLog.orcidid == orcidid)
+                ).order_by(ClaimsLog.id.desc()).first()
+                
+            if last_update is None:
+                q = session.query(ClaimsLog).filter_by(orcidid=orcidid).order_by(ClaimsLog.id.asc())
+            else:
+                if get_date(last_update.created) == updt:
+                    if force:
+                        self.logger.info("Profile {0} unchanged, but forced update in effect.".format(orcidid))
+                    else:
+                        self.logger.info("Skipping {0} (profile unchanged)".format(orcidid))
+                        return
+                q = session.query(ClaimsLog).filter(
+                    and_(ClaimsLog.orcidid == orcidid, ClaimsLog.id > last_update.id)) \
+                    .order_by(ClaimsLog.id.asc())
+                        
             
+            # now get info about each record #TODO: enhance the matching (and refactor)
+            # we'll try to match identifiers against our own API; if a document is found
+            # it will be added to the `orcid_present` with corresponding timestamp (cdate)
+            orcid_present = {}
+            for w in works:
+                bibc = None
+                try:
+                    ids =  w['work-external-identifiers']['work-external-identifier']
+                    seek_ids = []
+                    
+                    # painstakingly check ids (start from a bibcode) if we can find it
+                    # we'll send it through (but start from bibcodes, then dois, arxiv...)
+                    fmap = orcid_identifiers_order
+                    for x in ids:
+                        xtype = x.get('work-external-identifier-type', None)
+                        if xtype:
+                            seek_ids.append((fmap.get(xtype.lower().strip(), fmap.get('*', -1)), 
+                                             x['work-external-identifier-id']['value']))
+                    
+                    if len(seek_ids) == 0:
+                        continue
+                    
+                    seek_ids = sorted(seek_ids, key=lambda x: x[0], reverse=True)
+                    for _priority, fvalue in seek_ids:
+                        try:
+                            time.sleep(1.0/random.randint(1, 10)) # be nice to the api
+                            metadata = self.retrieve_metadata(fvalue, search_identifiers=True)
+                            if metadata and metadata.get('bibcode', None):
+                                bibc = metadata.get('bibcode')
+                                self.logger.info('Match found {0} -> {1}'.format(fvalue, bibc))
+                                break
+                        except Exception, e:
+                            self.logger.error('Exception while searching for matching bibcode for: {}'.format(fvalue))
+                            self.logger.warning(e.message)
+                            
+                    
+                    if bibc:
+                        # would you believe that orcid doesn't return floats?
+                        ts = str(w['last-modified-date']['value'])
+                        ts = float('%s.%s' % (ts[0:10], ts[10:]))
+                        ts = datetime.datetime.fromtimestamp(ts, tzutc())
+                        try:
+                            provenance = w['source']['source-name']['value']
+                        except KeyError:
+                            provenance = 'orcid-profile'
+                        orcid_present[bibc.lower().strip()] = (bibc.strip(), get_date(ts.isoformat()), provenance)
+                    else:
+                        self.logger.warning('Found no bibcode for: {orcidid}. {ids}'.format(ids=json.dumps(ids), orcidid=orcidid))
+                        
+                except KeyError, e:
+                    self.logger.warning('Error processing a record: '
+                        '{0} ({1})'.format(w,
+                                           traceback.format_exc()))
+                    continue
+                except TypeError, e:
+                    self.logger.warning('Error processing a record: '
+                        '{0} ({1})'.format(w,
+                                           traceback.format_exc()))
+                    continue
+    
             
-            if r.text.strip() == "":
-                return task_check_orcid_profiles.apply_async(msg, countdown = total_wait)
+            # find all records we have processed at some point
+            updated = {}
+            removed = {}
             
-            data = r.json()
+            for cl in q.all():
+                if not cl.bibcode:
+                    continue
+                bibc = cl.bibcode.lower()
+                if cl.status == 'removed':
+                    removed[bibc] = (cl.bibcode, get_date(cl.created))
+                    if bibc in updated:
+                        del updated[bibc]
+                else: #elif cl.status in ('claimed', 'updated', 'forced', 'unchanged'):
+                    updated[bibc] = (cl.bibcode, get_date(cl.created))
+                    if bibc in removed:
+                        del removed[bibc]
             
-            if len(data) == 0:
-                return task_check_orcid_profiles.apply_async(msg, countdown = total_wait)
-            
-            msg['errcount'] = 0 # success, we got data from the api, reset the counter
-
-            # we received the data, immediately update the databaes (so that other processes don't 
-            # ask for the same starting date)            
-            # data should be ordered by date updated (but to be sure, let's check it); we'll save it
-            # as latest 'check point'
-            dates = [utils.get_date(x['updated']) for x in data]
-            dates = sorted(dates, reverse=True)
-            
-            kv.value = dates[0].isoformat()
-            session.merge(kv)
+            return orcid_present, updated, removed
+        
+    
+        
+    @cachetools.cached(cache)  
+    def retrieve_orcid(self, orcid):
+        """
+        Finds (or creates and returns) model of ORCID
+        from the dbase. It will automatically update our
+        knowledge about the author every time it gets
+        called.
+        
+        :param orcid - String (orcid id)
+        :return - OrcidModel datastructure
+        """
+        with self.session_scope() as session:
+            u = session.query(AuthorInfo).filter_by(orcidid=orcid).first()
+            if u is not None:
+                return self.update_author(u)
+            u = self.create_orcid(orcid)
+            session.add(u)
             session.commit()
             
-            for rec in data:
-                payload = {'orcidid': rec['orcid_id'], 'start': latest_point.isoformat()}
-                task_index_orcid_profile.delay(payload)
+            return session.query(AuthorInfo).filter_by(orcidid=orcid).first().toJSON()
     
+    @cachetools.cached(orcid_cache)
+    def get_public_orcid_profile(self, orcidid):
+        r = requests.get(self._config.get('API_ORCID_PROFILE_ENDPOINT') % orcidid,
+                     headers={'Accept': 'application/json'})
+        if r.status_code != 200:
+            return None
+        else:
+            return r.json()
+    
+    @cachetools.cached(ads_cache)
+    def get_ads_orcid_profile(self, orcidid):
+        r = requests.get(self._config.get('API_ORCID_EXPORT_PROFILE') % orcidid,
+                     headers={'Accept': 'application/json', 'Authorization': 'Bearer:%s' % self._config.get('API_TOKEN')})
+        if r.status_code != 200:
+            return None
+        else:
+            return r.json()
+    
+    
+    def update_author(self, author):
+        """Updates existing AuthorInfo records. 
+        
+        It will check for new information. If there is a difference,
+        updates the record and also records the old values.
+        
+        :param: author - AuthorInfo instance
+        
+        :return: AuthorInfo object
+        
+        :sideeffect: Will insert new records (ChangeLog) and also update
+         the author instance
+        """
+        try:
+            new_facts = self.harvest_author_info(author.orcidid)
+        except:
+            return author.toJSON()
+        
+        info = author.toJSON()
+        with self.session_scope() as session:
+            old_facts = info['facts']
+            attrs = set(new_facts.keys())
+            attrs = attrs.union(old_facts.keys())
+            is_dirty = False
+            
+            for attname in attrs:
+                if old_facts.get(attname, None) != new_facts.get(attname, None):
+                    session.add(ChangeLog(key=u'{0}:update:{1}'.format(author.orcidid, attname), 
+                               oldvalue=json.dumps(old_facts.get(attname, None)),
+                               newvalue=json.dumps(new_facts.get(attname, None))))
+                    is_dirty = True
+            
+            if bool(author.account_id) != bool(new_facts.get('authorized', False)):
+                author.account_id = new_facts.get('authorized', False) and 1 or None 
+            
+            if is_dirty:
+                author.facts = json.dumps(new_facts)
+                author.name = new_facts.get('name', author.name)
+                aid=author.id
+                session.commit()
+                return session.query(AuthorInfo).filter_by(id=aid).first().toJSON()
+            else:
+                return info
 
 
-if __name__ == '__main__':
-    app.start()
+    def create_orcid(self, orcid, name=None, facts=None):
+        """
+        Creates an ORCID object and populates it with data
+        (this endpoint will query the API to discover
+        information about the author; so it is potentially
+        expensive)
+        
+        :param: orcid - String, ORCID ID
+        :param: name - String, name of the author (optional)
+        :param: facts - dictionary of other facts we want to
+            know/store (about the author)
+        
+        :return: AuthorInfo object
+        """
+        name = names.cleanup_name(name)
+        
+        # retrieve profile from our own orcid microservice
+        if not name or not facts:
+            profile = self.harvest_author_info(orcid, name, facts)
+            name = name or profile.get('name', None)
+            if not name:
+                raise IgnorableException('Cant find an author name for orcid-id: {}'.format(orcid))
+            facts = profile
+    
+        return AuthorInfo(orcidid=orcid, name=name, facts=json.dumps(facts), account_id=facts.get('authorized', None) and 1 or None)
+    
+    
+    def harvest_author_info(self, orcidid, name=None, facts=None):
+        """
+        Does the hard job of querying public and private 
+        API's for whatever information we want to collect
+        about the ORCID ID;
+        
+        At this stage, we want to mainly retrieve author
+        names (ie. variations of the author name)
+        
+        :param: orcidid - String
+        :param: name - String, name of the author (optional)
+        :param: facts - dict, info about the author
+        
+        :return: dict with various keys: name, author, author_norm, orcid_name
+                (if available)
+        """
+        
+        author_data = {}
+        
+        # first verify the public ORCID profile
+        j = self.get_public_orcid_profile(orcidid)
+        if j is None:
+            self.logger.error('We cant verify public profile of: http://orcid.org/%s' % orcidid)
+        else:
+            # we don't trust (the ugly) ORCID profiles too much
+            # j['orcid-profile']['orcid-bio']['personal-details']['family-name']
+            if 'orcid-profile' in j and 'orcid-bio' in j['orcid-profile'] \
+                and 'personal-details' in j['orcid-profile']['orcid-bio'] and \
+                'family-name' in j['orcid-profile']['orcid-bio']['personal-details'] and \
+                'given-names' in j['orcid-profile']['orcid-bio']['personal-details']:
+                
+                fname = (j['orcid-profile']['orcid-bio']['personal-details'].get('family-name', {}) or {}).get('value', None)
+                gname = (j['orcid-profile']['orcid-bio']['personal-details'].get('given-names', {}) or {}).get('value', None)
+                
+                if fname and gname:
+                    author_data['orcid_name'] = ['%s, %s' % (fname, gname)]
+                    author_data['name'] = author_data['orcid_name'][0]
+                
+                    
+        # search for the orcidid in our database (but only the publisher populated fiels)
+        # we can't trust other fiels to bootstrap our database
+        r = requests.get(
+                    '%(endpoint)s?q=%(query)s&fl=author,author_norm,orcid_pub&rows=100&sort=pubdate+desc' % \
+                    {
+                     'endpoint': self._config.get('API_SOLR_QUERY_ENDPOINT'),
+                     'query' : 'orcid_pub:%s' % self.cleanup_orcidid(orcidid),
+                    },
+                    headers={'Authorization': 'Bearer %s' % self._config.get('API_TOKEN')})
+        
+        if r.status_code != 200:
+            self.logger.error('Failed getting data from our own API! (err: %s)' % r.status_code)
+            raise Exception(r.text)
+        
+        
+        # go through the documents and collect all the names that correspond to the ORCID
+        master_set = {}
+        for doc in r.json()['response']['docs']:
+            for k,v in names.extract_names(orcidid, doc).items():
+                if v:
+                    master_set.setdefault(k, {})
+                    n = names.cleanup_name(v)
+                    if not master_set[k].has_key(n):
+                        master_set[k][n] = 0
+                    master_set[k][n] += 1
+        
+        # get ADS data about the user
+        # 0000-0003-3052-0819 | {"authorizedUser": true, "currentAffiliation": "Australian Astronomical Observatory", "nameVariations": ["Green, Andrew W.", "Green, Andy", "Green, Andy W."]}
+    
+        r = self.get_ads_orcid_profile(orcidid)
+        if r:
+            _author = r
+            _info = _author.get('info', {}) or {}
+            if _info.get('authorizedUser', False):
+                author_data['authorized'] = True
+            if _info.get('currentAffiliation', False):
+                author_data['current_affiliation'] = _info['currentAffiliation']
+            _vars = _info.get('nameVariations', None)
+            if _vars:
+                master_set.setdefault('author', {})
+                for x in _vars:
+                    x = names.cleanup_name(x)
+                    v = master_set['author'].get(x, 1)
+                    master_set['author'][x] = v
+        
+        # elect the most frequent name to become the 'author name'
+        # TODO: this will choose the normalized names (as that is shorter)
+        # maybe we should choose the longest (but it is not too important
+        # because the matcher will be checking all name variants during
+        # record update)
+        mx = 0
+        for k,v in master_set.items():
+            author_data[k] = sorted(list(v.keys()))
+            for name, freq in v.items():
+                if freq > mx:
+                    author_data['name'] = name
+        
+        # automatically add the short names, because they make us find
+        # more matches
+        short_names = set()
+        for x in ('author', 'orcid_name', 'author_norm'):
+            if x in author_data and author_data[x]:
+                for name in author_data[x]:
+                    for variant in names.build_short_forms(name):
+                        short_names.add(variant)
+        if len(short_names):
+            author_data['short_name'] = sorted(list(short_names))
+        
+        return author_data
+    
+    
+    @cachetools.cached(bibcode_cache) 
+    def retrieve_metadata(self, bibcode, search_identifiers=False):
+        """
+        From the API retrieve the set of metadata we want to know about the record.
+        """
+        params={
+                'q': search_identifiers and 'identifier:"{0}"'.format(bibcode) or 'bibcode:"{0}"'.format(bibcode),
+                'fl': 'author,bibcode,identifier'
+                }
+        r = requests.get(self._config.get('API_SOLR_QUERY_ENDPOINT'),
+             params=params,
+             headers={'Accept': 'application/json', 'Authorization': 'Bearer:%s' % self._config.get('API_TOKEN')})
+        if r.status_code != 200:
+            raise Exception('{}\n{}\n{}'.format(r.status_code, params, r.text))
+        else:
+            data = r.json().get('response', {})
+            if data.get('numFound') == 1:
+                docs = data.get('docs', [])
+                return docs[0]
+            elif data.get('numFound') == 0:
+                if search_identifiers:
+                    bibcode_cache.setdefault(bibcode, {}) # insert to prevent failed retrievals
+                    raise IgnorableException(u'No metadata found for identifier:{0}'.format(bibcode))
+                else:
+                    return self.retrieve_metadata(bibcode, search_identifiers=True)
+            else:
+                if data.get('numFound') > 10:
+                    raise IgnorableException(u'Insane num of results for {0} ({1})'.format(bibcode, data.get('numFound')))
+                docs = data.get('docs', [])
+                for d in docs:
+                    for ir in d.get('identifier', []):
+                        if ir.lower().strip() == bibcode.lower().strip():
+                            return d
+                raise IgnorableException(u'More than one document found for {0}'.format(bibcode))
+        
+    
+    
+    def retrieve_record(self, bibcode):
+        """
+        Gets a record from the database (creates one if necessary)
+        """
+        with self.session_scope() as session:
+            r = session.query(Records).filter_by(bibcode=bibcode).first()
+            if r is None:
+                r = Records(bibcode=bibcode)
+                session.add(r)
+            out = r.toJSON()
+            
+            metadata = self.retrieve_metadata(bibcode)
+            authors = metadata.get('author', [])
+            
+            if out.get('authors') != authors:
+                r.authors = json.dumps(authors)
+                out['authors'] = authors
+            
+            session.commit()
+            return out
+        
+        
+    def record_claims(self, bibcode, claims, authors=None):
+        """
+        Stores results of the processing in the database.
+        
+        :param: bibcode
+        :type: string
+        :param: claims
+        :type: dict
+        """
+        
+        if not isinstance(claims, basestring):
+            claims = json.dumps(claims)
+        if authors and not isinstance(authors, basestring):
+            authors = json.dumps(authors)
+            
+        with self.session_scope() as session:
+            if not isinstance(claims, basestring):
+                claims = json.dumps(claims)
+            r = session.query(Records).filter_by(bibcode=bibcode).first()
+            if r is None:
+                t = get_date()
+                r = Records(bibcode=bibcode, 
+                            claims=claims, 
+                            created=t,
+                            updated=t,
+                            authors=authors
+                            )
+                session.add(r)
+            else:
+                r.updated = get_date()
+                r.claims = claims
+                if authors:
+                    r.authors = authors
+                session.merge(r)
+            session.commit()
+
+
+    def mark_processed(self, bibcode):
+        """Updates the date on which the record has been processed (i.e.
+        something has consumed it
+        
+        :param: bibcode
+        :type: str
+        
+        :return: None
+        """
+        
+        with self.session_scope() as session:
+            r = session.query(Records).filter_by(bibcode=bibcode).first()
+            if r is None:
+                raise IgnorableException('Nonexistant record for {0}'.format(bibcode))
+            r.processed = get_date()
+            session.commit()
+            return True
