--- ADSOrcid/pipeline/exceptions.py
@@ -1,10 +1,0 @@
-
-class ProcessingException(Exception):
-    """Recoverable exception, should be reported to the
-    ErrorHandler."""
-    pass
-
-class IgnorableException(Exception):
-    """Dont mind restarting the worker."""
-    pass
-
