--- ADSOrcid/pipeline/workers.py
@@ -1,2 +1,0 @@
-from . import ClaimsIngester, ErrorHandler, OutputHandler, \
-  OrcidImporter, ClaimsRecorder
